Remove implicits
