+++ colossus/src/main/scala/colossus/protocols/http/package.scala
@@ -5,6 +5,7 @@ import colossus.metrics.TagMap
 
 import service._
 
+import scala.language.higherKinds
 
 package object http extends HttpBodyEncoders with HttpBodyDecoders {
 
@@ -25,33 +26,16 @@ package object http extends HttpBodyEncoders with HttpBodyDecoders {
     type Response = HttpResponse
   }
 
-  object Http extends ClientFactories[Http, HttpClient] {
+  object HttpClientLifter extends ClientLifter[Http, HttpClient] {
+    override def lift[M[_]](client: Sender[Http, M], clientConfig: Option[ClientConfig])(implicit async: Async[M]): HttpClient[M] = {
+      new BasicLiftedClient(client, clientConfig) with HttpClient[M]
+    }
+  }
+
+  object Http extends ClientFactories[Http, HttpClient](HttpClientLifter) {
 
     implicit lazy val clientFactory = ServiceClientFactory.basic("http", () => new StaticHttpClientCodec)
 
-    class ServerDefaults  {
-      def errorResponse(error: ProcessingFailure[HttpRequest]) = error match {
-        case RecoverableError(request, reason) => reason match {
-          case c: UnhandledRequestException => request.notFound(s"No route for ${request.head.url}")
-          case other => request.error(reason.toString)
-        }
-        case IrrecoverableError(reason) => {
-          HttpResponse(HttpResponseHead(HttpVersion.`1.1`, HttpCodes.BAD_REQUEST,  HttpHeaders.Empty), HttpBody("Bad Request"))
-        }
-      }
-    }
-
-    class ClientDefaults  {
-      def name = "http"
-    }
-
-    object defaults  {
-      
-      implicit val httpServerDefaults = new ServerDefaults
-
-      implicit val httpClientDefaults = new ClientDefaults
-
-    }
   }
 
   trait HttpMessage[H <: HttpMessageHead] extends BaseHttpMessage[H, HttpBody]
