+++ colossus/src/main/scala/colossus/protocols/memcache/package.scala
@@ -3,6 +3,8 @@ package protocols
 
 import service._
 
+import scala.language.higherKinds
+
 package object memcache {
 
   trait Memcache extends Protocol {
@@ -10,13 +12,15 @@ package object memcache {
     type Response = MemcacheReply
   }
 
-  object Memcache extends ClientFactories[Memcache, MemcacheClient] {
+  object MemcacheClientLifter extends ClientLifter[Memcache, MemcacheClient] {
+    override def lift[M[_]](client: Sender[Memcache, M], clientConfig: Option[ClientConfig])(implicit async: Async[M]): MemcacheClient[M] = {
+      new BasicLiftedClient(client, clientConfig) with MemcacheClient[M]
+    }
+  }
+
+  object Memcache extends ClientFactories[Memcache, MemcacheClient](MemcacheClientLifter) {
 
     implicit lazy val clientFactory = ServiceClientFactory.basic("memcache", () => new MemcacheClientCodec)
-
-    object defaults {
-      //???
-    }
 
   }
