+++ colossus/src/main/scala/colossus/protocols/redis/RedisOps.scala
@@ -495,17 +495,6 @@ trait RedisClient[M[_]] extends LiftedClient[Redis, M] {
 
 }
 
-object RedisClient {
-
-  implicit object RedisClientLifter extends ClientLifter[Redis, RedisClient] {
-
-    override def lift[M[_]](client: Sender[Redis, M], clientConfig: Option[ClientConfig])(implicit async: Async[M]): RedisClient[M] = {
-      new BasicLiftedClient(client, clientConfig) with RedisClient[M]
-    }
-  }
-
-}
-
 
 case class RedisException(message : String) extends Exception
